reorganize into library and parallel examples
