+++ .gitignore
@@ -1,3 +1,4 @@
 /bin/
-/node_modules/
+**/node_modules/
 Pulumi.*.yaml
+/.vscode/
