+++ examples/datawarehouse/Pulumi.yaml
