+++ examples/datawarehouse/package.json
@@ -1,0 +1,25 @@
+{
+  "name": "aws-typescript",
+  "devDependencies": {
+    "@babel/preset-typescript": "^7.7.2",
+    "@types/jest": "^24.0.23",
+    "@types/node": "^8.0.0",
+    "jest": "^24.9.0",
+    "ts-jest": "^24.1.0",
+    "typescript": "^3.7.2"
+  },
+  "dependencies": {
+    "@pulumi/aws": "^1.0.0",
+    "@pulumi/awsx": "^0.18.10",
+    "@pulumi/pulumi": "^1.0.0",
+    "@types/moment-timezone": "^0.5.12",
+    "athena-client": "^2.5.1",
+    "aws-sdk": "^2.569.0",
+    "moment-timezone": "^0.5.27",
+    "uuid": "^3.3.3"
+  },
+  "scripts": {
+    "test": "jest",
+    "test-watch": "jest --watch"
+  }
+}
