+++ examples/datawarehouse/put_record.sh
