+++ examples/datawarehouse/template.Pulumi.serverless-db.yaml
