+++ tsconfig.json
@@ -13,6 +13,6 @@
         "forceConsistentCasingInFileNames": true
     },
     "files": [
-        "index.ts"
+        "examples/datawarehouse/index.ts"
     ]
 }
